feat/frontend: Renamed some files and added common constant file
