+++ frontend/src/core/AppType.ts
