+++ frontend/src/core/CoreConstants.ts
@@ -1,0 +1,2 @@
+export const maxFileSizePerFile = 1000000 * 20; //20 mb
+export const maxTotalFileSize = 1000000 * 20 * 5; //100 mb
